refactor: 💡 clean up old registration scripts and styles
