+++ src/components/overlay/ts/handlePasswordResetOverlay.ts
@@ -1,4 +1,4 @@
-import { redirectToApp } from '../../registrationFormular/ts/autoLogin';
+import { redirectToApp } from '../../registration/ts/autoLogin';
 
 export const initPasswordResetOverlay = () => {
 	const passwordResetOverlay = document.querySelector(
