+++ src/components/registration/ts/autoLogin.ts
