--- src/components/registrationFormular/scss/_registrationFormular.styles.scss
@@ -1,58 +1,0 @@
-// TODO: CHECK IF STYLES STILL NEEDED WITH NEW REGISTRATION
-
-.registration {
-	.formWrapper__infoText {
-		&__text {
-			&--tertiary {
-				color: $tertiary;
-
-				.suggestionWrapper__item__content__teamAgency__icon svg {
-					margin-top: 0 !important;
-
-					&* {
-						fill: $tertiary;
-					}
-				}
-			}
-		}
-	}
-
-	.voluntaryInformation {
-		width: 100%;
-
-		&__content {
-			&__row:not(:last-child) {
-				margin-bottom: $grid-base-five;
-			}
-		}
-
-		.multiselect {
-			max-width: 690px;
-		}
-	}
-
-	.registration__form--u25 {
-		.select__wrapper {
-			min-width: 290px;
-			margin: 12px 0 $grid-base-three 26px;
-		}
-
-		#ageSelect__wrapper {
-			.custom-select-panel {
-				z-index: 10;
-			}
-		}
-
-		#stateSelect__wrapper {
-			.custom-select-container {
-				&:before {
-					z-index: 2;
-				}
-
-				&:after {
-					z-index: 1;
-				}
-			}
-		}
-	}
-}
