+++ src/resources/scss/styles.scss
@@ -52,7 +52,6 @@
 @import '../../components/profile/scss/profile.styles';
 @import '../../components/radioButton/scss/radiobutton.styles';
 @import '../../components/registration/scss/registration.styles';
-@import '../../components/registrationFormular/scss/registrationFormular.styles';
 @import '../../components/session/scss/session.styles';
 @import '../../components/session/scss/session.yellowTheme.styles';
 @import '../../components/sessionsList/scss/sessionsList.styles';
